+++ tests/Integration/Services/IM/IMServiceBuilderTest.php
@@ -1,0 +1,48 @@
+<?php
+
+/**
+ * This file is part of the bitrix24-php-sdk package.
+ *
+ * © Maksim Mesilov <mesilov.maxim@gmail.com>
+ *
+ * For the full copyright and license information, please view the MIT-LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Tests\Integration\Services\IM;
+
+use Bitrix24\SDK\Core\Exceptions\BaseException;
+use Bitrix24\SDK\Core\Exceptions\TransportException;
+use Bitrix24\SDK\Services\IM\IMServiceBuilder;
+use Bitrix24\SDK\Services\IM\Notify\Service\Notify;
+use Bitrix24\SDK\Services\ServiceBuilder;
+use Bitrix24\SDK\Tests\Integration\Factory;
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\Test;
+use PHPUnit\Framework\Attributes\TestDox;
+use PHPUnit\Framework\TestCase;
+
+#[CoversClass(IMServiceBuilder::class)]
+class IMServiceBuilderTest extends TestCase
+{
+    private ServiceBuilder $sb;
+
+    #[Test]
+    #[TestDox('Test send notification from system')]
+    public function testFooo(): void
+    {
+        var_dump($this->sb->getPlacementScope()->placement()->bind(
+            $this->sb->getIMScope()->placementLocationCodes()::IM_TEXTAREA,
+            ''
+        ));
+    }
+
+
+    #[\Override]
+    protected function setUp(): void
+    {
+        $this->sb = Factory::getServiceBuilder(true);
+    }
+}