+++ tests/Unit/Services/IM/Placements/ContextMenuPlacementOptionsTest.php
@@ -15,8 +15,8 @@ namespace Bitrix24\SDK\Tests\Unit\Services\IM\Placements;
 
 use Bitrix24\SDK\Services\IM\Placements\ChatContext;
 use Bitrix24\SDK\Services\IM\Placements\ContextMenuPlacementOptions;
-use Bitrix24\SDK\Services\IM\Placements\ExtranetAvailability;
-use Bitrix24\SDK\Services\IM\Placements\Role;
+use Bitrix24\SDK\Services\Placement\ExtranetAvailability;
+use Bitrix24\SDK\Services\Placement\Role;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\Attributes\TestDox;
@@ -37,7 +37,7 @@ class ContextMenuPlacementOptionsTest extends TestCase
     public function testFullPayload(): void
     {
         $contextMenuPlacementOptions = (new ContextMenuPlacementOptions())
-            ->context(ChatContext::User, ChatContext::Chat)
+            ->context(ChatContext::USER, ChatContext::CHAT)
             ->role(Role::User)
             ->extranet(ExtranetAvailability::Denied);
 
@@ -57,7 +57,7 @@ class ContextMenuPlacementOptionsTest extends TestCase
     {
         $contextMenuPlacementOptions = new ContextMenuPlacementOptions();
 
-        $this->assertSame($contextMenuPlacementOptions, $contextMenuPlacementOptions->context(ChatContext::All));
+        $this->assertSame($contextMenuPlacementOptions, $contextMenuPlacementOptions->context(ChatContext::ALL));
         $this->assertSame($contextMenuPlacementOptions, $contextMenuPlacementOptions->role(Role::Admin));
         $this->assertSame($contextMenuPlacementOptions, $contextMenuPlacementOptions->extranet(ExtranetAvailability::Allowed));
     }